requirements.txt
fastapi
python-dotenv
SQLAlchemy
uvicorn
psycopg2
psycopg2-binary
pydantic
